smart_complaint/pyproject.toml
[project]
name = "smart-complaint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "openai==1.79.0",
    "streamlit>=1.20.0",
    "altair==4.2.2",
    "textblob==0.17.1",
    "vaderSentiment==3.3.2",
    "python-dotenv==0.21.1",
    "uvicorn==0.22.0",
    "pytest==7.4.4"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.workspace]
members = [
    "smart_complaint",
]
